atualização

+++ src/ecommerce/repository/ProdutoRepository.java
@@ -10,7 +10,7 @@ public interface ProdutoRepository {
 		public void cadastrar(Produto produto);
 		public void atualizar(Produto produto);
 		public void deletar(int numero);
-		public void comprar(int numero, float quantidade);
+		public void comprar(int numero, int quantidade);
 		
 
 }
